Moving PPSTools/Utilities to Utilities/PPS as advised and the related changes in the code that depends on it
