+++ FastSimulation/CTPPSRecHitProducer/plugins/CTPPSRecHitProducer.cc
@@ -45,7 +45,7 @@ Implementation:
 
 #include "FastSimDataFormats/CTPPSFastSim/interface/CTPPSFastRecHitContainer.h"
 #include "FastSimulation/CTPPSFastGeometry/interface/CTPPSToFDetector.h"
-#include "PPSTools/Utilities/interface/PPSUnitConversion.h"
+#include "Utilities/PPS/interface/PPSUnitConversion.h"
 
 #include "DataFormats/GeometryVector/interface/LocalPoint.h"
 #include "DataFormats/GeometryVector/interface/LocalVector.h"
