+++ SimTransport/PPSProtonTransport/BuildFile.xml
@@ -3,12 +3,12 @@
 <use   name="CommonTools/Utils"/>
 <use   name="SimDataFormats/GeneratorProducts"/>
 <use   name="SimGeneral/HepPDTRecord"/>
-<use   name="PPSTools/Utilities"/>
+<use   name="Utilities/PPS"/>
 <use   name="SimTransport/TotemRPProtonTransportParametrization"/>
 <use   name="hector"/>
 <use   name="hepmc"/>
 <use   name="clhep"/>
-<use   name="root"/>
+<use   name="rootcore"/>
 <export>
   <lib name="1"/>
 </export>
