+++ Utilities/PPS/bin/BuildFile.xml
@@ -1,5 +1,5 @@
 <bin   file="PPSOpticsCalibration.cc" name="PPSOpticsCalibration">
-<use name="PPSTools/Utilities"/>
+<use name="Utilities/PPS"/>
 <use name="hector"/>
 <use name="hepmc"/>
 <use name="boost_program_options"/>
