+++ Utilities/PPS/bin/PPSOpticsCalibration.cc
@@ -1,4 +1,5 @@
-#include "PPSTools/Utilities/interface/PPSOpticsCalibrator.h"
+#include "Utilities/PPS/interface/PPSOpticsCalibrator.h"
+#include "Utilities/PPS/interface/PPSUnitConversion.h"
 #include <TEllipse.h>
 #include <iomanip>
 //using namespace std;
